+++ src/CompositeRepository.php
@@ -12,6 +12,8 @@
 namespace Puli\Repository;
 
 use Puli\Repository\Resource\Collection\ResourceCollection;
+use Puli\Repository\Resource\Collection\ResourceCollectionInterface;
+use Puli\Repository\Resource\ResourceInterface;
 use Puli\Repository\Uri\RepositoryFactoryException;
 use Webmozart\PathUtil\Path;
 
@@ -69,6 +71,11 @@ class CompositeRepository implements ResourceRepositoryInterface
     private $repos = array();
 
     /**
+     * @var string[][]
+     */
+    private $shadows = array();
+
+    /**
      * Mounts a repository to a path.
      *
      * The repository may either be passed as {@link ResourceRepositoryInterface}
@@ -111,11 +118,16 @@ class CompositeRepository implements ResourceRepositoryInterface
             ));
         }
 
-        if ('/' === $path) {
-            throw new InvalidPathException('The root "/" cannot be mounted.');
-        }
+        $path = Path::canonicalize($path);
 
-        $this->repos[Path::canonicalize($path)] = $repositoryFactory;
+        $this->repos[$path] = $repositoryFactory;
+        $this->shadows[$path] = array();
+
+        // Prefer more specific mount points (e.g. "/app) over less specific
+        // ones (e.g. "/")
+        krsort($this->repos);
+
+        $this->rebuildShadows();
     }
 
     /**
@@ -148,7 +160,12 @@ class CompositeRepository implements ResourceRepositoryInterface
             ));
         }
 
-        unset($this->repos[Path::canonicalize($path)]);
+        $path = Path::canonicalize($path);
+
+        unset($this->repos[$path]);
+        unset($this->shadows[$path]);
+
+        $this->rebuildShadows();
     }
 
     /**
@@ -156,7 +173,7 @@ class CompositeRepository implements ResourceRepositoryInterface
      */
     public function get($path)
     {
-        list ($mountPoint, $subPath) = $this->parsePath($path);
+        list ($mountPoint, $subPath) = $this->splitPath($path);
 
         if (null === $mountPoint) {
             throw new ResourceNotFoundException(sprintf(
@@ -165,7 +182,9 @@ class CompositeRepository implements ResourceRepositoryInterface
             ));
         }
 
-        return $this->getRepository($mountPoint)->get($subPath);
+        $resource = $this->getRepository($mountPoint)->get($subPath);
+
+        return '/' === $mountPoint ? $resource : $resource->createReference($path);
     }
 
     /**
@@ -173,13 +192,16 @@ class CompositeRepository implements ResourceRepositoryInterface
      */
     public function find($selector)
     {
-        list ($mountPoint, $subSelector) = $this->parsePath($selector);
+        list ($mountPoint, $subSelector) = $this->splitPath($selector);
 
         if (null === $mountPoint) {
             return new ResourceCollection();
         }
 
-        return $this->getRepository($mountPoint)->find($subSelector);
+        $resources = $this->getRepository($mountPoint)->find($subSelector);
+        $this->replaceByReferences($resources, $mountPoint);
+
+        return $resources;
     }
 
     /**
@@ -187,7 +209,7 @@ class CompositeRepository implements ResourceRepositoryInterface
      */
     public function contains($selector)
     {
-        list ($mountPoint, $subSelector) = $this->parsePath($selector);
+        list ($mountPoint, $subSelector) = $this->splitPath($selector);
 
         if (null === $mountPoint) {
             return false;
@@ -201,7 +223,7 @@ class CompositeRepository implements ResourceRepositoryInterface
      */
     public function listDirectory($path)
     {
-        list ($mountPoint, $subPath) = $this->parsePath($path);
+        list ($mountPoint, $subPath) = $this->splitPath($path);
 
         if (null === $mountPoint) {
             throw new ResourceNotFoundException(sprintf(
@@ -210,7 +232,10 @@ class CompositeRepository implements ResourceRepositoryInterface
             ));
         }
 
-        return $this->getRepository($mountPoint)->listDirectory($subPath);
+        $resources = $this->getRepository($mountPoint)->listDirectory($subPath);
+        $this->replaceByReferences($resources, $mountPoint);
+
+        return $resources;
     }
 
     /**
@@ -229,17 +254,27 @@ class CompositeRepository implements ResourceRepositoryInterface
             ));
         }
 
-        $collection = new ResourceCollection();
+        $resources = new ResourceCollection();
 
-        foreach ($this->repos as $repo) {
+        // The repositories are sorted in reverse order. Do a reverse traversal
+        // to add resources in sorted order
+        for (end($this->repos); null !== key($this->repos); prev($this->repos)) {
+            $repo = current($this->repos);
+            $mountPoint = key($this->repos);
+
             try {
-                $collection->merge($repo->findByTag($tag));
+                $tagged = $repo->findByTag($tag);
             } catch (ResourceNotFoundException $e) {
-                // continue
+                continue;
             }
+
+            $this->filterOvershadowedResources($tagged, $mountPoint);
+            $this->replaceByReferences($tagged, $mountPoint);
+
+            $resources->merge($tagged);
         }
 
-        return $collection;
+        return $resources;
     }
 
     /**
@@ -257,11 +292,22 @@ class CompositeRepository implements ResourceRepositoryInterface
             }
         }
 
-        // reindex
-        return array_values(array_unique($tags));
+        $tags = array_unique($tags);
+
+        sort($tags);
+
+        return $tags;
     }
 
-    private function parsePath($path)
+    /**
+     * Splits a path into mount point and path.
+     *
+     * @param string $path The path to split.
+     *
+     * @return array An array with the mount point and the path. If no mount
+     *               point was found, both are `null`.
+     */
+    private function splitPath($path)
     {
         if ('' === $path) {
             throw new InvalidPathException('The mount point must not be empty.');
@@ -285,6 +331,11 @@ class CompositeRepository implements ResourceRepositoryInterface
 
         foreach ($this->repos as $mountPoint => $_) {
             if (Path::isBasePath($mountPoint, $path)) {
+                // Special case "/": return the complete path
+                if ('/' === $mountPoint) {
+                    return array($mountPoint, $path);
+                }
+
                 return array($mountPoint, substr($path, strlen($mountPoint)));
             }
         }
@@ -323,5 +374,90 @@ class CompositeRepository implements ResourceRepositoryInterface
         }
 
         return $this->repos[$mountPoint];
+    }
+
+    /**
+     * Filters out overshadowed resources from a resource collection.
+     *
+     * Read {@link rebuildShadows()} to learn more about shadows.
+     *
+     * @param ResourceCollectionInterface $resources  The resources to filter.
+     * @param string                      $mountPoint The mount point from which
+     *                                                the resources were loaded.
+     */
+    private function filterOvershadowedResources(ResourceCollectionInterface $resources, $mountPoint)
+    {
+        foreach ($resources as $key => $resource) {
+            $path = $resource->getPath();
+
+            foreach ($this->shadows[$mountPoint] as $shadow) {
+                if (Path::isBasePath($shadow, $path)) {
+                    unset($resources[$key]);
+                }
+            }
+        }
+    }
+
+    /**
+     * Replaces all resources in the collection by references.
+     *
+     * If a resource "/resource" was loaded from a mount point "/mount", the
+     * resource is replaced by a reference with the path "/mount/resource".
+     *
+     * @param ResourceCollectionInterface $resources  The resources to replace.
+     * @param string                      $mountPoint The mount point from which
+     *                                                the resources were loaded.
+     */
+    private function replaceByReferences(ResourceCollectionInterface $resources, $mountPoint)
+    {
+        if ('/' !== $mountPoint) {
+            foreach ($resources as $key => $resource) {
+                $resources[$key] = $resource->createReference($mountPoint.$resource->getPath());
+            }
+        }
+    }
+
+    /**
+     * Rebuilds the shadows of the mount points.
+     *
+     * This method should be called after adding or removing mount points.
+     *
+     * Shadows are mount points that are located within another mount point.
+     * For example, if the paths "/app" and "/app/data" are mounted, then
+     * "/data" is a shadow of "/app". This means that Resources in the "/data"
+     * directory are overshadowed by the resources in the "/app/data"
+     * repository.
+     *
+     * Overshadowed resources are filtered out from the results of all methods.
+     */
+    private function rebuildShadows()
+    {
+        $mountPoints = array_keys($this->shadows);
+
+        foreach ($mountPoints as $mountPoint) {
+            $this->shadows[$mountPoint] = array();
+
+            foreach ($mountPoints as $maybeShadow) {
+                if ($mountPoint === $maybeShadow) {
+                    continue;
+                }
+
+                // $mountPoint = "/app"
+                // $maybeShadow = "/app/data"
+                // => $maybeShadow is a shadow
+
+                // $mountPoint = "/app"
+                // $maybeShadow = "/data"
+                // => $maybeShadow is not a shadow
+
+                // The root "/" is overshadowed by all other mount points
+                if ('/' === $mountPoint) {
+                    $this->shadows[$mountPoint][] = $maybeShadow;
+                } elseif (Path::isBasePath($mountPoint, $maybeShadow)) {
+                    // Only store "/data" for the shadow "/app/data" in "/app"
+                    $this->shadows[$mountPoint][] = substr($maybeShadow, strlen($mountPoint));
+                }
+            }
+        }
     }
 }