+++ tests/CompositeRepositoryTest.php
@@ -45,7 +45,6 @@ class CompositeRepositoryTest extends \PHPUnit_Framework_TestCase
             array(''),
             array(null),
             array(123),
-            array('/'), // root cannot be mounted
             array('\\'),
             array(new \stdClass()),
         );
@@ -57,20 +56,59 @@ class CompositeRepositoryTest extends \PHPUnit_Framework_TestCase
     public function testMountRepository($mountPoint)
     {
         $repo = $this->getMock('Puli\Repository\ResourceRepositoryInterface');
+        $resource = new TestFile('/path/to/resource');
+        $resource->attachTo($repo);
+        $mountedPath = rtrim($mountPoint, '/').'/path/to/resource';
 
         $this->repo->mount($mountPoint, $repo);
 
         $repo->expects($this->once())
             ->method('get')
             ->with('/path/to/resource')
-            ->will($this->returnValue('RESULT'));
+            ->will($this->returnValue($resource));
 
-        $this->assertEquals('RESULT', $this->repo->get(rtrim($mountPoint, '/').'/path/to/resource'));
+        $expected = $resource->createReference($mountedPath);
+
+        $this->assertEquals($expected, $this->repo->get($mountedPath));
+    }
+
+    public function testGetPrefersMostSpecificMountPoint()
+    {
+        $repo1 = $this->getMock('Puli\Repository\ResourceRepositoryInterface');
+        $repo2 = $this->getMock('Puli\Repository\ResourceRepositoryInterface');
+        $resource1 = new TestFile('/resource1');
+        $resource1->attachTo($repo1);
+        $resource2 = new TestFile('/resource2');
+        $resource2->attachTo($repo2);
+
+        $this->repo->mount('/', $repo1);
+        $this->repo->mount('/app', $repo2);
+
+        $repo1->expects($this->once())
+            ->method('get')
+            ->with('/resource1')
+            ->will($this->returnValue($resource1));
+        $repo2->expects($this->once())
+            ->method('get')
+            ->with('/resource2')
+            ->will($this->returnValue($resource2));
+
+        $result1 = $this->repo->get('/resource1');
+
+        $this->assertSame($resource1, $result1);
+
+        $result2 = $this->repo->get('/app/resource2');
+
+        $this->assertEquals('/app/resource2', $result2->getPath());
+        $this->assertEquals('/resource2', $result2->getRepositoryPath());
+        $this->assertSame($repo2, $result2->getRepository());
     }
 
     public function testMountRepositoryFactory()
     {
         $repo = $this->getMock('Puli\Repository\ResourceRepositoryInterface');
+        $resource = new TestFile('/path/to/resource');
+        $resource->attachTo($repo);
 
         $this->repo->mount('/webmozart', function () use ($repo) {
             return $repo;
@@ -79,9 +117,11 @@ class CompositeRepositoryTest extends \PHPUnit_Framework_TestCase
         $repo->expects($this->once())
             ->method('get')
             ->with('/path/to/resource')
-            ->will($this->returnValue('RESULT'));
+            ->will($this->returnValue($resource));
 
-        $this->assertEquals('RESULT', $this->repo->get('/webmozart/path/to/resource'));
+        $expected = $resource->createReference('/webmozart/path/to/resource');
+
+        $this->assertEquals($expected, $this->repo->get('/webmozart/path/to/resource'));
     }
 
     /**
@@ -224,6 +264,27 @@ class CompositeRepositoryTest extends \PHPUnit_Framework_TestCase
         $this->assertFalse($this->repo->contains('/webmozart/path/to/resource-2'));
     }
 
+    public function testContainsPrefersMostSpecificMountPoint()
+    {
+        $repo1 = $this->getMock('Puli\Repository\ResourceRepositoryInterface');
+        $repo2 = $this->getMock('Puli\Repository\ResourceRepositoryInterface');
+
+        $this->repo->mount('/', $repo1);
+        $this->repo->mount('/app', $repo2);
+
+        $repo1->expects($this->once())
+            ->method('contains')
+            ->with('/resource-1')
+            ->will($this->returnValue(true));
+        $repo2->expects($this->once())
+            ->method('contains')
+            ->with('/resource-2')
+            ->will($this->returnValue(false));
+
+        $this->assertTrue($this->repo->contains('/resource-1'));
+        $this->assertFalse($this->repo->contains('/app/resource-2'));
+    }
+
     public function testContainsReturnsFalseIfMountPointNotFound()
     {
         $this->assertFalse($this->repo->contains('/webmozart'));
@@ -256,15 +317,51 @@ class CompositeRepositoryTest extends \PHPUnit_Framework_TestCase
     public function testFind()
     {
         $repo = $this->getMock('Puli\Repository\ResourceRepositoryInterface');
+        $resource1 = new TestFile('/path/to/res1');
+        $resource2 = new TestFile('/path/to/res2');
 
         $this->repo->mount('/webmozart', $repo);
 
         $repo->expects($this->once())
             ->method('find')
             ->with('/path/to/res*')
-            ->will($this->returnValue('RESULT'));
+            ->will($this->returnValue(new ResourceCollection(array(
+                $resource1,
+                $resource2,
+            ))));
 
-        $this->assertSame('RESULT', $this->repo->find('/webmozart/path/to/res*'));
+        $expected = new ResourceCollection(array(
+            $resource1->createReference('/webmozart/path/to/res1'),
+            $resource2->createReference('/webmozart/path/to/res2'),
+        ));
+
+        $this->assertEquals($expected, $this->repo->find('/webmozart/path/to/res*'));
+    }
+
+    public function testFindPrefersMostSpecificMountPoint()
+    {
+        $repo1 = $this->getMock('Puli\Repository\ResourceRepositoryInterface');
+        $repo2 = $this->getMock('Puli\Repository\ResourceRepositoryInterface');
+        $resource1 = new TestFile('/res1');
+        $resource2 = new TestFile('/res2');
+
+        $this->repo->mount('/', $repo1);
+        $this->repo->mount('/app', $repo2);
+
+        $repo1->expects($this->once())
+            ->method('find')
+            ->with('/res1*')
+            ->will($this->returnValue(new ResourceCollection(array($resource1))));
+        $repo2->expects($this->once())
+            ->method('find')
+            ->with('/res2*')
+            ->will($this->returnValue(new ResourceCollection(array($resource2))));
+
+        $expected1 = new ResourceCollection(array($resource1));
+        $expected2 = new ResourceCollection(array($resource2->createReference('/app/res2')));
+
+        $this->assertEquals($expected1, $this->repo->find('/res1*'));
+        $this->assertEquals($expected2, $this->repo->find('/app/res2*'));
     }
 
     public function testFindReturnsEmptyCollectionIfMountPointNotFound()
@@ -302,15 +399,49 @@ class CompositeRepositoryTest extends \PHPUnit_Framework_TestCase
     public function testListDirectory()
     {
         $repo = $this->getMock('Puli\Repository\ResourceRepositoryInterface');
+        $resource1 = new TestFile('/path/to/dir/file1');
+        $resource2 = new TestFile('/path/to/dir/file2');
 
         $this->repo->mount('/webmozart', $repo);
 
         $repo->expects($this->once())
             ->method('listDirectory')
             ->with('/path/to/dir')
-            ->will($this->returnValue('RESULT'));
+            ->will($this->returnValue(new ResourceCollection(array(
+                $resource1,
+                $resource2,
+            ))));
 
-        $this->assertSame('RESULT', $this->repo->listDirectory('/webmozart/path/to/dir'));
+        $expected = new ResourceCollection(array(
+            $resource1->createReference('/webmozart/path/to/dir/file1'),
+            $resource2->createReference('/webmozart/path/to/dir/file2'),
+        ));
+
+        $this->assertEquals($expected, $this->repo->listDirectory('/webmozart/path/to/dir'));
+    }
+
+    public function testListRootDirectory()
+    {
+        $repo = $this->getMock('Puli\Repository\ResourceRepositoryInterface');
+        $resource1 = new TestFile('/path/to/dir/file1');
+        $resource2 = new TestFile('/path/to/dir/file2');
+
+        $this->repo->mount('/', $repo);
+
+        $repo->expects($this->once())
+            ->method('listDirectory')
+            ->with('/path/to/dir')
+            ->will($this->returnValue(new ResourceCollection(array(
+                $resource1,
+                $resource2,
+            ))));
+
+        $expected = new ResourceCollection(array(
+            $resource1,
+            $resource2,
+        ));
+
+        $this->assertEquals($expected, $this->repo->listDirectory('/path/to/dir'));
     }
 
     /**
@@ -349,28 +480,61 @@ class CompositeRepositoryTest extends \PHPUnit_Framework_TestCase
     {
         $repo1 = $this->getMock('Puli\Repository\ResourceRepositoryInterface');
         $repo2 = $this->getMock('Puli\Repository\ResourceRepositoryInterface');
+        $resource1 = new TestFile('/file1');
+        $resource2 = new TestFile('/file2');
 
         $this->repo->mount('/webmozart', $repo1);
         $this->repo->mount('/puli', $repo2);
 
-        $resources = new ResourceCollection(array(
-            new TestFile('foo'),
-            new TestFile('bar'),
+        $repo1->expects($this->once())
+            ->method('findByTag')
+            ->with('acme/tag')
+            ->will($this->returnValue(new ResourceCollection(array($resource1))));
+        $repo2->expects($this->once())
+            ->method('findByTag')
+            ->with('acme/tag')
+            ->will($this->returnValue(new ResourceCollection(array($resource2))));
+
+        // Result is sorted
+        $expected = new ResourceCollection(array(
+            $resource2->createReference('/puli/file2'),
+            $resource1->createReference('/webmozart/file1'),
         ));
+
+        $this->assertEquals($expected, $this->repo->findByTag('acme/tag'));
+    }
+
+    public function testFindByTagEliminatesOvershadowedTags()
+    {
+        $repo1 = $this->getMock('Puli\Repository\ResourceRepositoryInterface');
+        $repo2 = $this->getMock('Puli\Repository\ResourceRepositoryInterface');
+        $resource1 = new TestFile('/file1');
+        $resource2 = new TestFile('/puli/file2'); // overshadowed by other mount
+        $resource3 = new TestFile('/file3');
+
+        $this->repo->mount('/', $repo1);
+        $this->repo->mount('/puli', $repo2);
 
         $repo1->expects($this->once())
             ->method('findByTag')
             ->with('acme/tag')
-            ->will($this->returnValue(new ResourceCollection(array($resources[0]))));
+            ->will($this->returnValue(new ResourceCollection(array(
+                $resource1,
+                $resource2, // overshadowed by other mount
+            ))));
         $repo2->expects($this->once())
             ->method('findByTag')
             ->with('acme/tag')
-            ->will($this->returnValue(new ResourceCollection(array($resources[1]))));
+            ->will($this->returnValue(new ResourceCollection(array(
+                $resource3,
+            ))));
 
-        $this->assertEquals(
-            $resources,
-            $this->repo->findByTag('acme/tag')
-        );
+        $expected = new ResourceCollection(array(
+            $resource1->createReference('/file1'),
+            $resource3->createReference('/puli/file3'),
+        ));
+
+        $this->assertEquals($expected, $this->repo->findByTag('acme/tag'));
     }
 
     /**
@@ -389,7 +553,7 @@ class CompositeRepositoryTest extends \PHPUnit_Framework_TestCase
         $this->repo->findByTag(new \stdClass());
     }
 
-    public function testGetTagsReturnsUnionFromAllRepositories()
+    public function testGetTagsReturnsSortedUnionFromAllRepositories()
     {
         $repo1 = $this->getMock('Puli\Repository\ResourceRepositoryInterface');
         $repo2 = $this->getMock('Puli\Repository\ResourceRepositoryInterface');
@@ -399,13 +563,13 @@ class CompositeRepositoryTest extends \PHPUnit_Framework_TestCase
 
         $repo1->expects($this->once())
             ->method('getTags')
-            ->will($this->returnValue(array('foo')));
+            ->will($this->returnValue(array('foo', 'zong')));
         $repo2->expects($this->once())
             ->method('getTags')
             ->will($this->returnValue(array('foo', 'bar')));
 
         $this->assertEquals(
-            array('foo', 'bar'),
+            array('bar', 'foo', 'zong'),
             $this->repo->getTags()
         );
     }