+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Changelog
  * `ResourceInterface` now extends `\Serializable`
  * renamed `LocalResource` to `AbstractLocalResource`
  * added `AbstractResource`
+ * `CompositeRepository` now sets the correct path for returned resources
 
 * 1.0.0-alpha4 (2014-12-03)